Force board state resync in cases where a client may register a new spawn.
Existing rules that affect the stats of a piece are mostly resolved client-side. For the host's overridden piece stats to propagate to clients, the host needs to somehow sync its state with each client.

It is important to trigger this sync as close to the spawn event as possible, else the client may interact with the piece given non-actual stats, potentially causing desync issues.

Due to how heavy the resync is, it is undesirable to naively sync periodically (e.g., at the end of each action). However, because Demeo has only fine-grained and super-heavy-grained sync events, some overhead is to be expected.