Return raw GameObjects (not wrapped by an additional GameObject) when creating new buttons or text. (#166)
Returning wrapped objects by default allowed us to preserve the composition and position of objects.  For example, text is pre-shifted -0.2 along the Z-axis so that it appears slightly in front of non-text objects.

While this is helpful, it has made it difficult to modify the direct components of the button/text objects.  Instead, callers will now have to wrap the result themselves (helper wrapper `WrapObject` still available).