Perform some maintenance on rules. (#230)
* Make `_adjustments` a non-static variable.

Otherwise, it is changed by all instantiations of this rule.  By making it non-static, a copy a new `_adjustments` is created for each instance of the rule.  And only after `OnActivate()` is called is that rule-specific `_adjustments` value copied to the global `_globalAdjustments` field.

* Simplify expression.

* Use method which can be reused on both OnPostGameCreated and OnDeactive.

* Shorted code in `StatusEffectConfigRule` that handled updating.

Reduced the number of loops used. Used a guard class to reduce nesting by one level.

* Use method for AbilityBackstabAdjustedRule which can be reused on both OnPostGameCreated and OnDeactive, and saves exact previous value.