Update `RoomFinder` mod for compatibility with 2023-05-16 Demeo update.
There were numerous changes to Demeo's game-joining flow that `RoomFinder` needed to compensate for.

Demeo has moved to location at which cached rooms were available (previously `cachedRoomList` in `GameStateMachine`, now `roomList` in `LobbyMatchmakingController`).

As a consequence of the above change, we must also now capture a reference to the constructed `LobbyMatchmakingController` as part of the patching sequence.

Some resources that were once in `BowserCore.dll` have been moved to `BowserLegacyRuntime.dll` and `BowserCoreDataTypesRuntime.dll`, meaning those are now new dependencies.