Remove references to now-unneeded button handler for Hangouts. (#475)
There was a TODO mentioned in #470 for removing use of CommonModule.HangoutsButtonHandler.RegisterBowserButton(). Before the latest Demeo update, we had to capture reference to a special object in the game in order to make buttons clickable while in Hangouts.

As of the latest Demeo update, clickable buttons and achieved in a different way, meaning we can remove our use of it.