New rule and uses
Added PieceKeyholder rule which will give the keyholder 1 damage resist and 1 counter-attack damage in most games. If used in a PointGainRule game it will instead give a Locked buff that shows the value of using the key on the exit lock and the buff will count down for each turn it's not used to exit.