Update the `author` field to correctly reflect the team behind the develpment effort. (#492)
Originally, the `author` field was set so as to attribute the mods to
their corresponding GitHub repository. However, in practice, that field
is used more for display purposes. Thus, this PR updates the `author`
field so that it correctly reflects the folks behind the effort.

Should pass all build checks once
https://github.com/orendain/DemeoMods/pull/491 is merged.

Co-authored-by: The Gray Alien <103865052+TheGrayAlien@users.noreply.github.com>