Add automatic release drafting when tagging mods for release. (#524)
Adds automatic release drafting when tagging mods for release. When a
tag is created for a mod, a release is automatically drafted with the
changelog automatically filled it.

In order to improve tagging, this PR assumes a slight change how to
releases are tagged.


Instead of `<version>-<mod-name>` (e.g., `v1.0.0-roomfinder`), the files
in this PR look for `<mod-name>-<version>` (e.g., `roomfinder-v1.0.0`).

This way, tags are organized first by the mod they represent before
anything else.