Also renamed rule
The original method of using ProjectileHitSequence OnStarted can no longer function because it doesn't have a way to identify the target anymore since v1.39.

Found no way to prevent the stun other than to make the player invulnerable for the turn before the stun happens. The invulnerability affect fades when the player gets to move so it's not too OP or anything.

Renamed the rule because this has the potential to fix more than electrical attacks since it grabs all the data needed from DealDamage.